+++ src/k8-types/src/app/stateful.rs
