+++ src/k8-types/src/core/mod.rs
@@ -5,4 +5,3 @@ pub mod pod;
 pub mod secret;
 pub mod service;
 
-pub use k8_obj_metadata as metadata;
