+++ src/k8-types/src/core/plugin.rs
@@ -1,11 +1,11 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use k8_obj_metadata::Crd;
-use k8_obj_metadata::CrdNames;
-use k8_obj_metadata::DefaultHeader;
-use k8_obj_metadata::Spec;
-use k8_obj_metadata::Status;
+use crate::Crd;
+use crate::CrdNames;
+use crate::DefaultHeader;
+use crate::Spec;
+use crate::Status;
 
 const CREDENTIAL_API: Crd = Crd {
     group: "client.authentication.k8s.io",
